more converter consolidation
