+++ src/App/ChaosRecipeEnhancer.UI/Converters/BoolConverters.cs
@@ -150,3 +150,40 @@ public sealed class MultiBoolToBoolAndConverter : IMultiValueConverter
         throw new NotImplementedException();
     }
 }
+
+/// <summary>
+/// Converts an enum value to a boolean value based on a specified parameter.
+/// </summary>
+[ValueConversion(typeof(Enum), typeof(bool))]
+public class EnumBooleanConverter : IValueConverter
+{
+    /// <summary>
+    /// Converts an enum value to a boolean value.
+    /// </summary>
+    /// <param name="value">The enum value to convert.</param>
+    /// <param name="targetType">The target type of the conversion.</param>
+    /// <param name="parameter">The parameter to compare the enum value against.</param>
+    /// <param name="culture">The culture information.</param>
+    /// <returns>True if the enum value equals the specified parameter; otherwise, false.</returns>
+    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+        // Check if the enum value equals the specified parameter
+        return value.Equals(parameter);
+    }
+
+    /// <summary>
+    /// Converts a boolean value back to an enum value.
+    /// </summary>
+    /// <param name="value">The boolean value to convert.</param>
+    /// <param name="targetType">The target type of the conversion.</param>
+    /// <param name="parameter">The parameter to return if the boolean value is true.</param>
+    /// <param name="culture">The culture information.</param>
+    /// <returns>The specified parameter if the boolean value is true; otherwise, <see cref="Binding.DoNothing"/>.</returns>
+    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+        // Check if the boolean value is true
+        // If true, return the specified parameter
+        // If false, return Binding.DoNothing to indicate that no conversion should occur
+        return value.Equals(true) ? parameter : Binding.DoNothing;
+    }
+}
