--- src/App/ChaosRecipeEnhancer.UI/Converters/EnumBooleanConverter.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Globalization;
-using System.Windows.Data;
-
-namespace ChaosRecipeEnhancer.UI.Converters;
-
-public class EnumBooleanConverter : IValueConverter
-{
-    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-    {
-        return value.Equals(parameter);
-    }
-
-    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-    {
-        return value.Equals(true) ? parameter : Binding.DoNothing;
-    }
-}
